Refactor create proposal voting
